# swagger.yaml
openapi: 3.0.0
info:
  title: Scientific Publications API
  version: 1.0.0
  description: |
    A comprehensive backend API for analyzing scientific publications with NLP and ML capabilities.
    This system provides endpoints for managing publications, authors, citations, topics, and analytics.
    
    ## Features
    - Publication management with full-text search
    - Author profiles with collaboration networks
    - Citation analysis and relationship mapping
    - Topic modeling and trend analysis
    - Advanced analytics and insights
    - Data import/export capabilities
    - Secure authentication and authorization
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    `Authorization: Bearer <your-jwt-token>`
    
    ## Rate Limiting
    API endpoints are rate-limited to prevent abuse. Limits vary by endpoint type.
  contact:
    name: API Support
    email: support@scientific-publications.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:4000/api
    description: Development server
paths:
...
  /analytics/publications/trends:
    get:
      tags:
        - Analytics
      summary: Get publication trends over time
      description: Retrieves publication trends over time with various granularities
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [year, month, quarter]
            default: year
        - name: years
          in: query
          schema:
            type: string
          description: Comma-separated years to analyze
      responses:
        '200':
          description: Publication trends retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  trends:
                    type: array
                    items:
                      type: object
                      properties:
                        _id:
                          type: string
                        count:
                          type: integer
                        total_citations:
                          type: integer
                  period:
                    type: string
                  years:
                    type: array
                    items:
                      type: integer
  /analytics/citations/trends:
    get:
      tags:
        - Analytics
      summary: Get citation trends over time
      description: Retrieves citation trends analysis
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [year, month, quarter]
            default: year
        - name: years
          in: query
          schema:
            type: string
          description: Comma-separated years to analyze
      responses:
        '200':
          description: Citation trends retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  trends:
                    type: array
                    items:
                      type: object
                  period:
                    type: string
                  years:
                    type: array
                    items:
                      type: integer
  /analytics/topics/trends:
    get:
      tags:
        - Analytics
      summary: Get topic trends analysis
      description: Retrieves trending topics and category distribution
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 50
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [total_publications, total_citations, popularity_score]
      responses:
        '200':
          description: Topic trends retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  trending_topics:
                    type: array
                    items:
                      $ref: '#/components/schemas/Topic'
                  category_distribution:
                    type: array
                    items:
                      type: object
  /analytics/authors/productivity:
    get:
      tags:
        - Analytics
      summary: Get author productivity metrics
      description: Retrieves author productivity analysis
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: metric
          in: query
          schema:
            type: string
            enum: [total_publications, total_citations, h_index, publications_per_year]
      responses:
        '200':
          description: Author productivity retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  authors:
                    type: array
                    items:
                      $ref: '#/components/schemas/Author'
                  metric:
                    type: string
                  total:
                    type: integer
  /analytics/collaborations/network:
    get:
      tags:
        - Analytics
      summary: Get collaboration network analysis
      description: Retrieves author collaboration network data
      parameters:
        - name: minCollaborations
          in: query
          schema:
            type: integer
            default: 5
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Collaboration network data
          content:
            application/json:
              schema:
                type: object
                properties:
                  nodes:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                        collaboration_count:
                          type: integer
                  edges:
                    type: array
                    items:
                      type: object
                      properties:
                        source:
                          type: string
                        target:
                          type: string
                        weight:
                          type: integer
                        last_collaboration_year:
                          type: integer
                  metrics:
                    type: object
                    properties:
                      total_nodes:
                        type: integer
                      total_edges:
                        type: integer
                      min_collaborations:
                        type: integer
  /analytics/geographic/distribution:
    get:
      tags:
        - Analytics
      summary: Get geographic distribution analysis
      description: Retrieves geographic distribution of authors, publications, or citations
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [authors, publications, citations]
            default: authors
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 50
      responses:
        '200':
          description: Geographic distribution retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  distribution:
                    type: array
                    items:
                      type: object
                      properties:
                        country:
                          type: string
                        count:
                          type: integer
                        total_citations:
                          type: integer
                        avg_h_index:
                          type: number
  /analytics/venues/analysis:
    get:
      tags:
        - Analytics
      summary: Get venue analysis
      description: Retrieves analysis of journals and conferences
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [journals, conferences]
            default: journals
        - name: metric
          in: query
          schema:
            type: string
            enum: [publications, citations, impact]
            default: publications
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 50
      responses:
        '200':
          description: Venue analysis retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  metric:
                    type: string
                  venues:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        publication_count:
                          type: integer
                        total_citations:
                          type: integer
                        avg_citations:
                          type: number
                        unique_authors:
                          type: integer
  /analytics/citations/impact:
    get:
      tags:
        - Analytics
      summary: Get citation impact analysis
      description: Retrieves high-impact publications analysis
      parameters:
        - name: timeframe
          in: query
          schema:
            type: string
            enum: [1year, 3years, 5years, 10years]
            default: 5years
        - name: minCitations
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
      responses:
        '200':
          description: Citation impact analysis retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  timeframe:
                    type: string
                  min_citations:
                    type: integer
                  metrics:
                    type: object
                  top_contributing_authors:
                    type: array
                    items:
                      type: object
                  sample_publications:
                    type: array
                    items:
                      $ref: '#/components/schemas/Publication'
  /analytics/research/trends:
    get:
      tags:
        - Analytics
      summary: Get research trends and predictions
      description: Retrieves research trend analysis and predictions
      parameters:
        - name: years
          in: query
          schema:
            type: integer
            default: 5
            minimum: 1
            maximum: 20
        - name: topicCount
          in: query
          schema:
            type: integer
            default: 20
            minimum: 5
            maximum: 50
      responses:
        '200':
          description: Research trends retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  analysis_period:
                    type: object
                  topic_trends:
                    type: array
                    items:
                      type: object
                  emerging_topics:
                    type: array
                    items:
                      $ref: '#/components/schemas/Topic'
                  declining_topics:
                    type: array
                    items:
                      $ref: '#/components/schemas/Topic'

  # Topic Routes
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      description: Number of items per page
...
  schemas:
...
    Institution:
      type: object
      properties:
        name:
          type: string
        department:
          type: string
        position:
          type: string
        country:
          type: string
...
    Publication:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        abstract:
          type: string
        publication_year:
          type: integer
        doi:
          type: string
        authors:
          type: array
          items:
            $ref: '#/components/schemas/AuthorReference'
        journal:
          $ref: '#/components/schemas/Journal'
        citations_count:
          type: integer
        keywords:
          type: array
          items:
            $ref: '#/components/schemas/Keyword'
        created_at:
          type: string
          format: date-time
...
    AuthorReference:
      type: object
      properties:
        author_id:
          type: string
        name:
          type: string
        order:
          type: integer
        affiliation:
          type: string
    Journal:
      type: object
      properties:
        name:
          type: string
        volume:
          type: string
        issue:
          type: string
        pages:
          type: string
        issn:
          type: string
        publisher:
          type: string
    Keyword:
      type: object
      properties:
        term:
          type: string
        weight:
          type: number
        source:
          type: string
...
    Author:
      type: object
      properties:
        id:
          type: string
        full_name:
          type: string
        orcid:
          type: string
        h_index:
          type: integer
        total_citations:
          type: integer
        total_publications:
          type: integer
        affiliations:
          type: array
          items:
            $ref: '#/components/schemas/Institution'
        research_interests:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
...
    Topic:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        keywords:
          type: array
          items:
            $ref: '#/components/schemas/Keyword'
        total_publications:
          type: integer
        total_citations:
          type: integer
        popularity_score:
          type: number
        trend_direction:
          type: string
        created_at:
          type: string
          format: date-time
...